Update 1: Planning + some rough coding
Start time: 10:05
End time:11:23
In this first log, I planned out my approach to project Euler's problem nine, asking for the Pythagorean triple of 1000 and began to try a bruteforce method. I understand that my method will not be very refined, as it is just a general test to try and figure out what the answer is. I will try and enable multi- thread processing (soon) for the host as the processing time is quite high given my method.