Update Number 6: Finished Calculation outline
Start time: 8:45
End time: 9:08
Today, I finished the mathematical outline and have solved for the answer requested by project Euler, but still must wait for a long period of time for it to finish running. I will be distributing the task into other threads but with an identical process and different queues to run in parallels. I think that I will take the project one step further by creating a list of all the possible pythag triples to print at the end. The solution I calculated for was x: 200, y: 375 and z: 425.0, with a sum of 1000 (x^2+y^2=z^2).