Update 7:  Finishing up
Start time: 6:20
End time: 8:56
Today, I mainly worked on troubleshooting, replicating code and running the scripts to get time stamps. My original objective of finding the Pythagorean triple of 1000 was a bit too easy to calculate (even via brute force) and took less than 10 seconds. I figured that it would be more resource and time consuming to calculate the triple of 10,000. It is a jump from 1 million (individual) combinations to a total of over 100 million. On average, (on my desktop), it takes around 8 minutes to calculate the triple of 10,000 using 8 threads, whereas with two threads, it takes more than 11 minutes. Please note that your results will very likely vary, even on my desktop, with a 16 gb of ram + 4.2ghz cpu, the process was bottlenecked (by the ram). When running the program, the anaconda prompt will eventually update, but will not show completion until much later and unfortunately, I cannot figure out why. The script will save a file to your directory, where the last line should display the solution, but also every possible Pythagorean triple between 1 and 10,000. If the script continues adding to the txt file, please quit the process, as there is a substantial delay until it stops populating your ram and ends the task.